HY_23.02.09_PointAccountImpl 수정 중

+++ src/main/java/com/keduit/helloworld/repository/MemberRepository.java
@@ -8,16 +8,16 @@ import org.springframework.data.jpa.repository.Query;
 import com.keduit.helloworld.entity.Favorites;
 import com.keduit.helloworld.entity.Member;
 
-public interface MemberRepository extends JpaRepository<Member, Integer>{
+public interface MemberRepository extends JpaRepository<Member, Long>{
 
 //	/**맴버 PK하나로 하나 불러오기*/
 //	@Query(value = "select * from member where member_num= :num order by member_num", nativeQuery = true)
-//	String getMemberNum(Integer num);
+//	String getMemberNum(Long num);
 	
 	@Query(value = "select * from Favorites as f join Member m on member_num = f.bookmarked " 
 			+ "where f.bookmarker= :bookmarker order by favorites_num desc", nativeQuery = true)
 	/** 즐겨찾기에서 마커를 넣으면 맴버 아이디가 출력됨(즐겨찾기 한 사람들) */
-	List<Member> getMemberMarker(Integer bookmarker);
+	List<Member> getMemberMarker(Long bookmarker);
 	
 	
 }
